Resolves #108; Refactored code and fix slow summary card load; Addresses part of #104
