+++ src/callbacks/__init__.py
@@ -1,0 +1,4 @@
+# callbacks/__init__.py
+
+from .charts_callbacks import register_charts_callbacks
+from .filters_callbacks import register_filters_callbacks
