+++ src/components/layout.py
@@ -2,191 +2,34 @@
 from dash import Dash, dcc, html, Input, Output, dash_table
 import dash_bootstrap_components as dbc
 import dash_vega_components as dvc
-from src.data.data_import import get_dropdown_options
-from src.charts.engagement_chart import engagement_chart
+from src.components.filters import create_global_filters
 from src.components.get_summary_stats import get_summary_stats
-from src.charts.make_density_plot import make_density_plot
-from src.charts.ranking_chart import create_wordcloud 
-from src.charts.make_popularity_score import make_popularity_score
-from src.charts.install_chart import installs_chart
+from src.components.chart_components import (engagement_chart_component, 
+                                             density_plot_component, 
+                                             popularity_histogram_component, 
+                                             wordcloud_component)
+from src.components.footer_components import footer_component
 
-def create_global_filters(df):
-    return [
-        dbc.Col([
-            html.H1([
-                html.Img(src='assets/android-chrome-192x192.png', height='50px'),
-                " Playstore Apps Ads Analytics"
-            ], 
-            className="text-center mb-3",
-            style={
-                'padding': '5px',
-                'backgroundColor': 'white'
-            }),
-            html.Br(),
-
-            html.H5('Global controls'),
-            html.Br(),
-
-            dbc.Label("Select Category:"),
-            dcc.Dropdown(
-                id="category-filter",
-                options=get_dropdown_options(df, "Category"),
-                value=["All"],
-                multi=True,
-                maxHeight=200
-            ),
-            html.Br(),
-
-            dbc.Label("Rating Range:"),
-            dcc.RangeSlider(
-                id="rating-slider",
-                min=1,
-                max=5,
-                step=0.5,
-                marks={i: str(i) for i in range(1, 6)},
-                value=[1, 5],
-                updatemode='mouseup',
-                tooltip={"always_visible": True, "placement": "bottom"}
-            ),
-            html.Br(),
-
-            dbc.Label("Select App Type:"),
-            dcc.Dropdown(id="app-type-filter", 
-                         options=get_dropdown_options(df, "Type"),
-                         value=["All"], 
-                         multi=True),
-            html.Br(),
-
-            dbc.Label("Select Content Rating:"),
-            dcc.Dropdown(
-                id="content-rating-filter",
-                options=get_dropdown_options(df, "Content Rating"),
-                value=["All"], 
-                multi=True
-            ),
-            html.Br()
-        ],
-        className="filter-80",
-        style={
-            'background-color': 'white',
-            'padding': 10,
-            'border-radius': 3,
-            'height': '100%'
-        })
-    ]
 
 def create_layout(df):
     # Filters
     global_filters = create_global_filters(df)
 
-    # Make charts
-    install_chart = dbc.Card([
-        dbc.CardHeader('Top 10 App Categories by Installs', style={'fontWeight': 'bold',
-                                                                   "textAlign": "center",}),
-        dbc.CardBody(
-            dvc.Vega(
-            id="category-chart",
-            spec=installs_chart(df).to_dict(format="vega"),
-            style={"padding": "20px", 
-                   "justifyContent": "center",  
-                   "alignItems": "center", 
-                   'width': '100%',
-                   'height': '100%' 
-                   }
-            ))
-    ],
-    className="shadow-sm h-100")
+    # Summary cards
+    summary_stats = get_summary_stats(df)
 
-    make_engagement_chart = dbc.Card([
-        dbc.CardHeader('Reviews vs. Installs for Top Apps', 
-                       className="fw-bold text-center bg-light border-bottom border-secondary"),
-        dbc.CardBody(
-            dvc.Vega(
-            id="engagement-chart",
-            spec=engagement_chart(df, ["All"]).to_dict(format="vega"),
-            style={'width': '100%',
-                   'height': '100%' 
-                   }
-            ),
-            className="d-flex align-items-center justify-content-center p-0"
-            )
-    ],
-    className="shadow-sm h-100 border-0 rounded")
+    # Charts
+    popularity_histogram = popularity_histogram_component(df)
+    make_engagement_chart = engagement_chart_component(df)
+    density_plot = density_plot_component(df)
+    wordcloud_chart = wordcloud_component(df)
 
-
-    density_plot = dbc.Card(
-    [
-        dbc.CardHeader(
-            "Density Plot for Ratings",
-            className="fw-bold text-center bg-light border-bottom border-secondary"
-        ),
-        dbc.CardBody(
-            dvc.Vega(
-            id="density-plot",
-            spec=make_density_plot(df, ["All"]).to_dict(format="vega"),
-            style={'width': '100%',
-                   'height': '100%' 
-                   }
-        ),
-        className="d-flex align-items-center justify-content-center p-0"
-        )
-    ],
-    className="shadow-sm h-100 border-0 rounded")
-
-    popularity_histogram = dbc.Card(
-        [
-            dbc.CardHeader(
-                "Average Popularity Score by Categories",
-                className="fw-bold text-center bg-light border-bottom border-secondary"
-            ),
-            dbc.CardBody(
-                dvc.Vega(
-                    id="popularity-histogram",
-                    spec=make_popularity_score(df, ["All"]).to_dict(format="vega"),
-                    style={
-                        "width": "100%",
-                        "height": "100%" 
-                    }
-                ),
-                className="d-flex align-items-center justify-content-center p-0" 
-            )
-        ],
-        className="shadow-sm h-100 border-0 rounded"
-    )
-
-
-    wordcloud_component = dbc.Card([
-        dbc.CardHeader('Word Cloud of Top Apps', 
-                       className="fw-bold text-center bg-light border-bottom border-primary"),
-        dbc.CardBody(
-            dcc.Graph(
-                id="wordcloud",
-                figure=create_wordcloud(df, ["All"]),
-                config={"displayModeBar": False},
-                style={'width': '100%', 
-                       'height': '100%'}
-            ),
-            className="d-flex align-items-center justify-content-center p-0" 
-        )
-    ],
-    className="shadow-sm h-100 border-0 rounded")
-
-    # Make footer
-    footer = dbc.Row(dbc.Col([
-            html.Hr(),
-            html.H6('This dashboard helps advertisement companies identify the most promising Google Play Store apps for ad placements by analyzing app metrics such as user engagement and ratings', 
-                    className="text-center fw-light mb-4",
-                    style={"maxWidth": "60%", "margin": "auto", "whiteSpace": "normal", "wordWrap": "break-word"}
-                    ),
-            html.P("Project by: Quanhua Huang, Yeji Sohn, Lukman Lateef, Ismail (Husain) Bhinderwala", className="text-center fw-light"),
-            html.P(["GitHub Repository: ", html.A("Link to Repo", href="https://github.com/UBC-MDS/DSCI-532_2025_25_Ads-Analytics", target="_blank")], className="text-center"),
-            html.P("Last Updated: March 2025", className="text-center fw-light"),
-            html.P([html.A("Adwords icons created by Freepik - Flaticon", href="https://www.flaticon.com/free-icons/adwords", title="adwords icons")], className="text-center fw-light mt-3")
-        ], width=12, className="text-center mt-4"))
+    # Footers
+    footer = footer_component()
 
     return dbc.Container([
         dbc.Row([
-            dbc.Col([ 
+            dbc.Col([
                 *global_filters
             ], md=2),
 
@@ -195,36 +38,36 @@ def create_layout(df):
                     dbc.Col(
                         dbc.Card([
                             dbc.CardHeader(html.Span("Average Rating", className="fw-bold text-dark"),
-                                        className="text-center bg-light border-bottom border-3",
-                                        style={"borderImage": "linear-gradient(to right, #4285F4, #EA4335, #FBBC05, #34A853) 1"}),
-                            dbc.CardBody(html.H3(id="mean-rating", className="fw-bold", style={"color": "#34A853"}), 
-                                        className="d-flex align-items-center justify-content-center p-4"),
+                                           className="text-center bg-light border-bottom border-3",
+                                           style={"borderImage": "linear-gradient(to right, #4285F4, #EA4335, #FBBC05, #34A853) 1"}),
+                            dbc.CardBody(html.H3(summary_stats['mean_rating'], id="mean-rating", className="fw-bold", style={"color": "#34A853"}),
+                                         className="d-flex align-items-center justify-content-center p-4"),
                         ], className="text-center shadow-sm border-0 rounded mb-3"),
                         md=4
                     ),
                     dbc.Col(
                         dbc.Card([
                             dbc.CardHeader(html.Span("Average Reviews", className="fw-bold text-dark"),
-                                        className="text-center bg-light border-bottom border-3",
-                                        style={"borderImage": "linear-gradient(to right, #4285F4, #EA4335, #FBBC05, #34A853) 1"}),
-                            dbc.CardBody(html.H3(id="mean-reviews", className="fw-bold", style={"color": "#4285F4"}),  
-                                        className="d-flex align-items-center justify-content-center p-4"),
+                                           className="text-center bg-light border-bottom border-3",
+                                           style={"borderImage": "linear-gradient(to right, #4285F4, #EA4335, #FBBC05, #34A853) 1"}),
+                            dbc.CardBody(html.H3(summary_stats['mean_reviews'], id="mean-reviews", className="fw-bold", style={"color": "#4285F4"}),
+                                         className="d-flex align-items-center justify-content-center p-4"),
                         ], className="text-center shadow-sm border-0 rounded mb-3"),
                         md=4
                     ),
                     dbc.Col(
                         dbc.Card([
                             dbc.CardHeader(html.Span("Average Installs", className="fw-bold text-dark"),
-                                        className="text-center bg-light border-bottom border-3",
-                                        style={"borderImage": "linear-gradient(to right, #4285F4, #EA4335, #FBBC05, #34A853) 1"}),
-                            dbc.CardBody(html.H3(id="mean-installs", className="fw-bold", style={"color": "#FBBC05"}), 
-                                        className="d-flex align-items-center justify-content-center p-4"),
+                                           className="text-center bg-light border-bottom border-3",
+                                           style={"borderImage": "linear-gradient(to right, #4285F4, #EA4335, #FBBC05, #34A853) 1"}),
+                            dbc.CardBody(html.H3(summary_stats['mean_installs'], id="mean-installs", className="fw-bold", style={"color": "#FBBC05"}),
+                                         className="d-flex align-items-center justify-content-center p-4"),
                         ], className="text-center shadow-sm border-0 rounded mb-3"),
                         md=4
                     ),
-                ], 
-                className="mt-3 mb-4", 
-                justify="center"),
+                ],
+                    className="mt-3 mb-4",
+                    justify="center"),
 
                 dbc.Row([
                     dbc.Col(dbc.Card([
@@ -235,7 +78,7 @@ def create_layout(df):
                             ]),
                             dbc.Row([
                                 dbc.Col(density_plot, md=6),
-                                dbc.Col(wordcloud_component, md=6)
+                                dbc.Col(wordcloud_chart, md=6)
                             ])
                         ])
                     ], style={"backgroundColor": "#f8f9fa", "paddingTop": "10px", "paddingBottom": "10px"}), md=12)
@@ -247,4 +90,3 @@ def create_layout(df):
         # Footer
         footer
     ], fluid=True, style={"backgroundColor": "#add8e6"})
-
